Adds drive code

+++ include/subsystemheaders/drive.h
@@ -12,4 +12,4 @@ extern bool run_flywheel;
 void op_indexer();
 void op_flywheel();
 void op_drive();
-extern int flypower;
+extern double flypower;
